requirements.txt
streamlit
griptape[all]
duckduckgo_search
google-generativeai 